Move accumulator config to its own file
